Consolodated some helper classes and functions for testing.
